add token queue, but not do free yet

+++ util.h
@@ -1,6 +1,6 @@
 #ifndef __UTIL__
 #define __UTIL__
 
-void split(char ***arr, char *str, char *del, int *count);
+void split(char ***arr, char *str, char *del, size_t *count);
 int regular_match(char *str, char *reg_str);
 #endif
